formFeedback: align message styling
